requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-api-python-client
youtube-transcript-api
openai
supabase-py 